+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 
+* The methods `xor`, `impl`, `equiv`, and `combine` of `NFAs` have been removed since the product automaton constructions do not correspond to the respective language operations.
 * `SimpleTS#getSuccessors(S state, Iterable<? extends I> input)` has been removed. In order to traverse a (non-deterministic) transition system, use `TransitionSystem#powersetView` instead.
 * The `net.automatalib:automata-serialization-core` artifact has been dropped. Its contents are now part of the `net.automatalib:automata-api` artifact.
 * `TAFParseDiagnosticListener` has been removed. Reporting is now done via (slf4j) logging.